ok gan xong

+++ src/hook/product/useMutationDeleteProduct.js
@@ -3,13 +3,17 @@ import { useMutation, useQueryClient } from "react-query";
 import { toast } from "react-toastify";
 import { catchAsyn } from "@/utils/catchAsync";
 
-const deleteProduct = catchAsyn(async ({ infoAdmin, productId }) => {
+const deleteProduct = catchAsyn(async ({ infoAdmin, productId, image }) => {
    const config = {
       headers: {
          Authorization: `Bearer ${infoAdmin.token}`,
       },
    };
-
+   const { data: success } = await axios.delete(`/api/image/${image}`);
+   if (success.result !== "ok") {
+      console.log(success);
+      throw new Error("Upload product failed");
+   }
    await axios.delete(`/api/products/${productId}`, config);
 });
